% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelSelection.R
\name{cv_krr}
\alias{cv_krr}
\title{Choose best lambda via cross-validation.}
\usage{
cv_krr(x, y, lambda_index)
}
\arguments{
\item{x}{Quantitative explanatory variables (vector, matrix, or data.frame).}

\item{y}{Quantitative response variable.}

\item{lambda_index}{Vector of lambdas to scan for best model.}
}
\value{
Returns best lambda, MSE, and model.  Also returns a dataframe of
every lambda_index value and corresponding MSE.
}
\description{
A very crude method to find the optimal lambda for a krr (kernel ridge
regression) model.
}

