% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOC.boot.R
\name{DOC.boot}
\alias{DOC.boot}
\title{Bootstraps for loess, negative slope and fns}
\usage{
DOC.boot(do, R = 100, subr = NULL, mov.avg = 5, span = 0.2,
  degree = 1, family = "symmetric", iterations = 4,
  surface = "interpolate", cores = 1)
}
\arguments{
\item{do}{Output from \code{DOC.do} function}

\item{R}{Number of bootstraps}

\item{subr}{If NULL will do bootstrap, alternatively an integer denoting size of subsample}

\item{mov.avg}{Moving average window to use for estimating where negative slope starts}

\item{span}{Span of loess smoothing}

\item{degree}{Degree of loess smoothing (If 1 linear, >1 polynomial)}

\item{family}{"gaussian" is least-squares fitting, "symmetric" is robust fitting}

\item{iterations}{Number of iterations for robust fitting}

\item{surface}{"direct" estimation (slow exact) or "interpolate" estimation (fast approximate)}

\item{cores}{Number of cores to use}
}
\value{
Output: A list with:
\itemize{
  \item A dataframe of loess for each bootstrap,
  \item a numeric with estimated slope from lme on data subsetted to Overlap values above median for each bootstrap,
  \item a numeric with estimated point of where negative slope starts (moving average of loess used) for each bootstrap,
  \item and a numeric with estimated Fns value for each bootstrap
}
}
\description{
Bootstraps for loess, negative slope and fns
}
