% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOC.ci.R
\name{DOC.ci}
\alias{DOC.ci}
\title{Get bootstrap Confidence Intervals}
\usage{
DOC.ci(boots, ci = c(0.025, 0.5, 0.975))
}
\arguments{
\item{boots}{Output from \code{DOC.boot} function}

\item{ci}{Vector with quantiles}
}
\value{
A dataframe with quantiles of bootstraped loess
}
\description{
Get bootstrap Confidence Intervals
}
