% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOC.do.R
\name{DOC.do}
\alias{DOC.do}
\title{Dissimilarity and Overlap}
\usage{
DOC.do(otu)
}
\arguments{
\item{otu}{Normalized otu-table (taxa as rows and samples sum to 1)}
}
\value{
A list with; a matrix of Overlap, a matrix of rJSD (Dissimilarities), and a dataframe with Overlap and rJSD as vectors
}
\description{
Dissimilarity and Overlap
}
