% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOC.loess.R
\name{DOC.loess}
\alias{DOC.loess}
\title{Calculate loess on the real data}
\usage{
DOC.loess(do, span = 0.2, degree = 1, family = "symmetric",
  iterations = 4, surface = "interpolate")
}
\arguments{
\item{do}{Output from DO function}

\item{span}{Span of loess smoothing}

\item{degree}{Degree of loess smoothing (If 1 linear, >1 polynomial)}

\item{family}{"gaussian" is least-squares fitting, "symmetric" is robust fitting}

\item{iterations}{Number of iterations for robust fitting}

\item{surface}{"direct" estimation (slow exact) or "interpolate" estimation (fast approximate)}
}
\value{
A dataframe with loess results from the non-bootstrapped data
}
\description{
Calculate loess on the real data
}
