% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOC.merge.R
\name{DOC.merge}
\alias{DOC.merge}
\title{Merge DOC objects and add a Name column, for example for plotting multiple DOCs}
\usage{
DOC.merge(x)
}
\arguments{
\item{x}{A named list with outputs from \code{DOC} function.}
}
\value{
A DOC object
}
\description{
Merge DOC objects and add a Name column, for example for plotting multiple DOCs
}
