% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOC.null.R
\name{DOC.null}
\alias{DOC.null}
\title{Wrapper to run a DOC analysis with null models}
\usage{
DOC.null(otu, N = 1, non.zero = TRUE, ...)
}
\arguments{
\item{otu}{An OTU-table (taxa as rows)}

\item{N}{Number of null models. Is 1 by default as in the orignal Nature paper.}

\item{non.zero}{Shuffle only non-zero (TRUE) or all values (FALSE)}

\item{...}{Arguments for the \code{DOC} function}
}
\value{
A list with results from \code{DOC} function. One element for each N
}
\description{
Wrapper to run a DOC analysis with null models
}
