% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOC.otunull.R
\name{DOC.otunull}
\alias{DOC.otunull}
\title{Make null matrix}
\usage{
DOC.otunull(otu, non.zero = TRUE)
}
\arguments{
\item{otu}{OTU-table with taxa as rows}

\item{non.zero}{Shuffle only non-zero (TRUE) or all values (FALSE)}
}
\value{
Shuffled OTU-table
}
\description{
Make null matrix
}
