% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOC.R
\name{DOC}
\alias{DOC}
\title{A wrapper to run the whole DOC analysis}
\usage{
DOC(otu, R = 100, subr = NULL, mov.avg = 5, ci = c(0.025, 0.5, 0.975),
  span = 0.2, degree = 1, family = "symmetric", iterations = 4,
  surface = "interpolate", cores = 1)
}
\arguments{
\item{otu}{An OTU-table (taxa as rows)}

\item{R}{Number of bootstraps}

\item{subr}{If NULL will do bootstrap, alternatively an integer denoting size of subsample}

\item{mov.avg}{Moving average window to use for estimating where negative slope starts}

\item{ci}{Vector with quantiles for confidence intervals}

\item{span}{Span of loess smoothing}

\item{degree}{Degree of loess smoothing (If 1 linear, >1 polynomial)}

\item{family}{"gaussian" is least-squares fitting, "symmetric" is robust fitting}

\item{iterations}{Number of iterations for robust fitting}

\item{surface}{"direct" estimation (slow exact) or "interpolate" estimation (fast approximate)}

\item{cores}{Number of cores to use}
}
\value{
Output: A list with class DOC:
\itemize{
  \item DO: A dataframe with Overlap and Dissimilarity (rJSD)
  \item LME: A dataframe with estimates of negative slope
  \item LOWESS: A dataframe with loess(lowess) results of non-bootstraped data
  \item NEG: A dataframe with Overlap values where negative slope starts
  \item FNS: A dataframe with Fns values
  \item BOOT: A dataframe with loess results for each bootstrap
  \item CI: A dataframe with quantiles for the BOOT dataframe
}
}
\description{
A wrapper to run the whole DOC analysis
}
