% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOC.rjsd.R
\name{DOC.rjsd}
\alias{DOC.rjsd}
\title{root Jensen-Shannon divergence}
\usage{
DOC.rjsd(x, y)
}
\arguments{
\item{x, y}{Numeric vectors}
}
\value{
A numeric
}
\description{
root Jensen-Shannon divergence between two numeric vectors
}
\keyword{dissimilarity}
\keyword{divergence}
\keyword{rJSD}
