% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classMethods.R
\name{fitted.krr}
\alias{fitted.krr}
\title{Produce fitted values}
\usage{
\method{fitted}{krr}(krrm)
}
\arguments{
\item{krrm}{A model of type krr.}
}
\value{
Fitted values
}
\description{
Produce fitted values
}

