% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createModelObject.R
\name{krr}
\alias{krr}
\title{Kernel Ridge Regression}
\usage{
krr(x, y, lambda, sigma = 1)
}
\arguments{
\item{x}{Quantitative explanatory variables (vector, matrix, or data.frame).}

\item{y}{Quantitative response variable.}

\item{lambda}{Tuning parameter for regression.}

\item{sigma}{Tuning parameter for Gaussian kernel.}
}
\value{
Returns the predictions, alphas, and MSE associated with the
regression, as well as the inputs x, y, the kernel function, and lambda.
}
\description{
A function for producing a krr (kernel ridge regression) model.
}

