% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.DOC.R
\name{plot.DOC}
\alias{plot.DOC}
\title{Plot results from DOC analysis}
\usage{
\method{plot}{DOC}(x, ...)
}
\arguments{
\item{x}{Output from \code{DOC} function}

\item{...}{Unused argument}
}
\description{
Plot results from DOC analysis
}
