% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classMethods.R
\name{plot.krr}
\alias{plot.krr}
\title{Produce fitted value plot.}
\usage{
\method{plot}{krr}(krrm)
}
\arguments{
\item{krrm}{A model of type krr.}
}
\value{
An aesthetic plot.
}
\description{
Produce fitted value plot.
}

