% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classMethods.R
\name{resid.krr}
\alias{resid.krr}
\title{Produce residuals}
\usage{
\method{resid}{krr}(krrm)
}
\arguments{
\item{krrm}{A model of type krr.}
}
\value{
Residuals
}
\description{
Produce residuals
}

